Create AbstractExpressiveContainerConfigTest with required test traits composed
This patch builds on [a comment on #2](https://github.com/zendframework/zend-container-config-test/pull/2#issuecomment-378847288).

First, it renames `ContainerTest` to `AbstractContainerTest`, in order
to follow our naming guidelines for abstract classes.

Second, it introduces `ExpressiveTestTrait`, which composes the various
traits that are _required_ by implementations in order to be compatible
with Expressive configuration.

Third, it modifies `AllTestTrait` to use `ExpressiveTestTrait` as well
as `SharedTestTrait`.

Fourth, it introduces `AbstractExpressiveContainerConfigTest`, which
extends `AbstractContainerTest` and composes `ExpressiveTestTrait`.

Fifth, it removes the `BaseContainerTest` as redundant to the new
`AbstractExpressiveContainerConfigTest`, and updates all test classes to
extend it.